# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "用户管理 CRUD 示例 API，演示 Go + Fiber + Ent + fx 分层架构",
        "title": "Go Fiber API Demo",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/health": {
            "get": {
                "description": "检查数据库和 Redis 连接状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "健康检查"
                ],
                "summary": "健康检查",
                "responses": {
                    "200": {
                        "description": "服务正常",
                        "schema": {
                            "$ref": "#/definitions/health.SwaggerHealthResponse"
                        }
                    },
                    "503": {
                        "description": "服务不可用",
                        "schema": {
                            "$ref": "#/definitions/health.SwaggerHealthResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "分页查询用户列表，支持关键词搜索",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "分页获取用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerUserListResponse"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go_sample_code_internal_handler_user.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerUserResponse"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/email/{email}": {
            "get": {
                "description": "根据邮箱地址获取用户详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "根据邮箱获取用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱地址",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerUserResponse"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/username/{username}": {
            "get": {
                "description": "根据用户名获取用户详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "根据用户名获取用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerUserResponse"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "根据用户 ID 获取用户详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "根据 ID 获取用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerUserResponse"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "根据用户 ID 更新用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go_sample_code_internal_handler_user.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerUserResponse"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据用户 ID 删除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerDeleteResponse"
                        }
                    },
                    "400": {
                        "description": "参数校验失败",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/user.SwaggerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "database.HealthData": {
            "type": "object",
            "properties": {
                "redis": {
                    "$ref": "#/definitions/database.RedisStatus"
                },
                "relational": {
                    "$ref": "#/definitions/database.RelationalStatus"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "database.RedisStatus": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "database.RelationalStatus": {
            "type": "object",
            "properties": {
                "driver": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "go_sample_code_internal_handler_user.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "nickname": {
                    "type": "string",
                    "maxLength": 64
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "phone": {
                    "type": "string",
                    "maxLength": 32
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "go_sample_code_internal_handler_user.ListUsersResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go_sample_code_internal_handler_user.UserResponse"
                    }
                }
            }
        },
        "go_sample_code_internal_handler_user.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "nickname": {
                    "type": "string",
                    "maxLength": 64
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "phone": {
                    "type": "string",
                    "maxLength": 32
                }
            }
        },
        "go_sample_code_internal_handler_user.UserResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "health.SwaggerHealthResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/database.HealthData"
                },
                "message": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "user.DeleteMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "user deleted successfully"
                }
            }
        },
        "user.SwaggerDeleteResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/user.DeleteMessage"
                },
                "message": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "user.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1003
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Request Validate Error"
                }
            }
        },
        "user.SwaggerUserListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/go_sample_code_internal_handler_user.ListUsersResponse"
                },
                "message": {
                    "type": "string",
                    "example": "OK"
                }
            }
        },
        "user.SwaggerUserResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/go_sample_code_internal_handler_user.UserResponse"
                },
                "message": {
                    "type": "string",
                    "example": "OK"
                }
            }
        }
    }
}